# backend/api/gen/go/admin/service/v1/i_product_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: admin/service/v1/i_product.proto

package adminpb

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/tx7do/go-crud/api/gen/go/pagination/v1"
	v11 "go-wind-shop/api/gen/go/catalog/service/v1"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationProductServiceBatchCreate = "/admin.service.v1.ProductService/BatchCreate"
const OperationProductServiceCount = "/admin.service.v1.ProductService/Count"
const OperationProductServiceCreate = "/admin.service.v1.ProductService/Create"
const OperationProductServiceCreateTranslation = "/admin.service.v1.ProductService/CreateTranslation"
const OperationProductServiceDelete = "/admin.service.v1.ProductService/Delete"
const OperationProductServiceDeleteTranslation = "/admin.service.v1.ProductService/DeleteTranslation"
const OperationProductServiceGet = "/admin.service.v1.ProductService/Get"
const OperationProductServiceGetTranslation = "/admin.service.v1.ProductService/GetTranslation"
const OperationProductServiceList = "/admin.service.v1.ProductService/List"
const OperationProductServiceTranslationExists = "/admin.service.v1.ProductService/TranslationExists"
const OperationProductServiceUpdate = "/admin.service.v1.ProductService/Update"
const OperationProductServiceUpdateTranslation = "/admin.service.v1.ProductService/UpdateTranslation"

type ProductServiceHTTPServer interface {
	// BatchCreate 批量创建商品
	BatchCreate(context.Context, *v11.BatchCreateProductsRequest) (*emptypb.Empty, error)
	// Count 统计商品数量
	Count(context.Context, *v1.PagingRequest) (*v11.CountProductResponse, error)
	// Create 创建商品
	Create(context.Context, *v11.CreateProductRequest) (*emptypb.Empty, error)
	// CreateTranslation 创建翻译
	CreateTranslation(context.Context, *v11.CreateProductTranslationRequest) (*v11.ProductTranslation, error)
	// Delete 删除商品
	Delete(context.Context, *v11.DeleteProductRequest) (*emptypb.Empty, error)
	// DeleteTranslation 删除翻译
	DeleteTranslation(context.Context, *v11.DeleteProductTranslationRequest) (*emptypb.Empty, error)
	// Get 查询商品详情
	Get(context.Context, *v11.GetProductRequest) (*v11.Product, error)
	// GetTranslation 获取翻译数据
	GetTranslation(context.Context, *v11.GetProductRequest) (*v11.ProductTranslation, error)
	// List 分页查询商品列表
	List(context.Context, *v1.PagingRequest) (*v11.ListProductResponse, error)
	// TranslationExists 检查翻译是否存在
	TranslationExists(context.Context, *v11.ProductTranslationExistsRequest) (*v11.ProductTranslationExistsResponse, error)
	// Update 更新商品
	Update(context.Context, *v11.UpdateProductRequest) (*emptypb.Empty, error)
	// UpdateTranslation 更新翻译
	UpdateTranslation(context.Context, *v11.UpdateProductTranslationRequest) (*v11.ProductTranslation, error)
}

func RegisterProductServiceHTTPServer(s *http.Server, srv ProductServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/admin/v1/mall/products", _ProductService_List27_HTTP_Handler(srv))
	r.GET("/admin/v1/mall/products/count", _ProductService_Count11_HTTP_Handler(srv))
	r.GET("/admin/v1/mall/products/{id}", _ProductService_Get27_HTTP_Handler(srv))
	r.POST("/admin/v1/mall/products", _ProductService_Create20_HTTP_Handler(srv))
	r.POST("/admin/v1/mall/products/batch", _ProductService_BatchCreate6_HTTP_Handler(srv))
	r.PUT("/admin/v1/mall/products/{id}", _ProductService_Update20_HTTP_Handler(srv))
	r.DELETE("/admin/v1/mall/products", _ProductService_Delete21_HTTP_Handler(srv))
	r.GET("/admin/v1/mall/products/{product_id}/translations/{language_code}", _ProductService_TranslationExists2_HTTP_Handler(srv))
	r.GET("/admin/v1/mall/products/{id}/translations", _ProductService_GetTranslation2_HTTP_Handler(srv))
	r.POST("/admin/v1/mall/products/translations", _ProductService_CreateTranslation2_HTTP_Handler(srv))
	r.PUT("/admin/v1/mall/products/{id}/translations", _ProductService_UpdateTranslation2_HTTP_Handler(srv))
	r.DELETE("/admin/v1/mall/products/{id}/translations", _ProductService_DeleteTranslation2_HTTP_Handler(srv))
}

func _ProductService_List27_HTTP_Handler(srv ProductServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListProductResponse)
		return ctx.Result(200, reply)
	}
}

func _ProductService_Count11_HTTP_Handler(srv ProductServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductServiceCount)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Count(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.CountProductResponse)
		return ctx.Result(200, reply)
	}
}

func _ProductService_Get27_HTTP_Handler(srv ProductServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetProductRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*v11.GetProductRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Product)
		return ctx.Result(200, reply)
	}
}

func _ProductService_Create20_HTTP_Handler(srv ProductServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.CreateProductRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductServiceCreate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Create(ctx, req.(*v11.CreateProductRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _ProductService_BatchCreate6_HTTP_Handler(srv ProductServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.BatchCreateProductsRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductServiceBatchCreate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.BatchCreate(ctx, req.(*v11.BatchCreateProductsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _ProductService_Update20_HTTP_Handler(srv ProductServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.UpdateProductRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductServiceUpdate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Update(ctx, req.(*v11.UpdateProductRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _ProductService_Delete21_HTTP_Handler(srv ProductServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.DeleteProductRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductServiceDelete)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Delete(ctx, req.(*v11.DeleteProductRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _ProductService_TranslationExists2_HTTP_Handler(srv ProductServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.ProductTranslationExistsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductServiceTranslationExists)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.TranslationExists(ctx, req.(*v11.ProductTranslationExistsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ProductTranslationExistsResponse)
		return ctx.Result(200, reply)
	}
}

func _ProductService_GetTranslation2_HTTP_Handler(srv ProductServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetProductRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductServiceGetTranslation)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetTranslation(ctx, req.(*v11.GetProductRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ProductTranslation)
		return ctx.Result(200, reply)
	}
}

func _ProductService_CreateTranslation2_HTTP_Handler(srv ProductServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.CreateProductTranslationRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductServiceCreateTranslation)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateTranslation(ctx, req.(*v11.CreateProductTranslationRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ProductTranslation)
		return ctx.Result(200, reply)
	}
}

func _ProductService_UpdateTranslation2_HTTP_Handler(srv ProductServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.UpdateProductTranslationRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductServiceUpdateTranslation)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateTranslation(ctx, req.(*v11.UpdateProductTranslationRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ProductTranslation)
		return ctx.Result(200, reply)
	}
}

func _ProductService_DeleteTranslation2_HTTP_Handler(srv ProductServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.DeleteProductTranslationRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductServiceDeleteTranslation)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteTranslation(ctx, req.(*v11.DeleteProductTranslationRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

type ProductServiceHTTPClient interface {
	// BatchCreate 批量创建商品
	BatchCreate(ctx context.Context, req *v11.BatchCreateProductsRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// Count 统计商品数量
	Count(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.CountProductResponse, err error)
	// Create 创建商品
	Create(ctx context.Context, req *v11.CreateProductRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// CreateTranslation 创建翻译
	CreateTranslation(ctx context.Context, req *v11.CreateProductTranslationRequest, opts ...http.CallOption) (rsp *v11.ProductTranslation, err error)
	// Delete 删除商品
	Delete(ctx context.Context, req *v11.DeleteProductRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// DeleteTranslation 删除翻译
	DeleteTranslation(ctx context.Context, req *v11.DeleteProductTranslationRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// Get 查询商品详情
	Get(ctx context.Context, req *v11.GetProductRequest, opts ...http.CallOption) (rsp *v11.Product, err error)
	// GetTranslation 获取翻译数据
	GetTranslation(ctx context.Context, req *v11.GetProductRequest, opts ...http.CallOption) (rsp *v11.ProductTranslation, err error)
	// List 分页查询商品列表
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.ListProductResponse, err error)
	// TranslationExists 检查翻译是否存在
	TranslationExists(ctx context.Context, req *v11.ProductTranslationExistsRequest, opts ...http.CallOption) (rsp *v11.ProductTranslationExistsResponse, err error)
	// Update 更新商品
	Update(ctx context.Context, req *v11.UpdateProductRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// UpdateTranslation 更新翻译
	UpdateTranslation(ctx context.Context, req *v11.UpdateProductTranslationRequest, opts ...http.CallOption) (rsp *v11.ProductTranslation, err error)
}

type ProductServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewProductServiceHTTPClient(client *http.Client) ProductServiceHTTPClient {
	return &ProductServiceHTTPClientImpl{client}
}

// BatchCreate 批量创建商品
func (c *ProductServiceHTTPClientImpl) BatchCreate(ctx context.Context, in *v11.BatchCreateProductsRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/mall/products/batch"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationProductServiceBatchCreate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Count 统计商品数量
func (c *ProductServiceHTTPClientImpl) Count(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.CountProductResponse, error) {
	var out v11.CountProductResponse
	pattern := "/admin/v1/mall/products/count"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationProductServiceCount))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Create 创建商品
func (c *ProductServiceHTTPClientImpl) Create(ctx context.Context, in *v11.CreateProductRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/mall/products"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationProductServiceCreate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// CreateTranslation 创建翻译
func (c *ProductServiceHTTPClientImpl) CreateTranslation(ctx context.Context, in *v11.CreateProductTranslationRequest, opts ...http.CallOption) (*v11.ProductTranslation, error) {
	var out v11.ProductTranslation
	pattern := "/admin/v1/mall/products/translations"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationProductServiceCreateTranslation))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Delete 删除商品
func (c *ProductServiceHTTPClientImpl) Delete(ctx context.Context, in *v11.DeleteProductRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/mall/products"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationProductServiceDelete))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// DeleteTranslation 删除翻译
func (c *ProductServiceHTTPClientImpl) DeleteTranslation(ctx context.Context, in *v11.DeleteProductTranslationRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/mall/products/{id}/translations"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationProductServiceDeleteTranslation))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Get 查询商品详情
func (c *ProductServiceHTTPClientImpl) Get(ctx context.Context, in *v11.GetProductRequest, opts ...http.CallOption) (*v11.Product, error) {
	var out v11.Product
	pattern := "/admin/v1/mall/products/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationProductServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetTranslation 获取翻译数据
func (c *ProductServiceHTTPClientImpl) GetTranslation(ctx context.Context, in *v11.GetProductRequest, opts ...http.CallOption) (*v11.ProductTranslation, error) {
	var out v11.ProductTranslation
	pattern := "/admin/v1/mall/products/{id}/translations"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationProductServiceGetTranslation))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// List 分页查询商品列表
func (c *ProductServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.ListProductResponse, error) {
	var out v11.ListProductResponse
	pattern := "/admin/v1/mall/products"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationProductServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// TranslationExists 检查翻译是否存在
func (c *ProductServiceHTTPClientImpl) TranslationExists(ctx context.Context, in *v11.ProductTranslationExistsRequest, opts ...http.CallOption) (*v11.ProductTranslationExistsResponse, error) {
	var out v11.ProductTranslationExistsResponse
	pattern := "/admin/v1/mall/products/{product_id}/translations/{language_code}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationProductServiceTranslationExists))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Update 更新商品
func (c *ProductServiceHTTPClientImpl) Update(ctx context.Context, in *v11.UpdateProductRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/mall/products/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationProductServiceUpdate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateTranslation 更新翻译
func (c *ProductServiceHTTPClientImpl) UpdateTranslation(ctx context.Context, in *v11.UpdateProductTranslationRequest, opts ...http.CallOption) (*v11.ProductTranslation, error) {
	var out v11.ProductTranslation
	pattern := "/admin/v1/mall/products/{id}/translations"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationProductServiceUpdateTranslation))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}